fix: split relation tests and import helpers from source
Address review: move relation tests to describe-relations.test.ts,
import buildRelationsOverview/printRelationsOverview from source
instead of reimplementing.
