+++ packages/cli/__tests__/describe.test.ts
@@ -93,24 +93,6 @@ const testRelation = defineRelation({
   toName: "orders",
 });
 
-const testRelation2 = defineRelation({
-  name: "order_customer",
-  from: "order",
-  to: "customer",
-  cardinality: "many_to_one",
-  fromName: "customer",
-  toName: "placed_orders",
-});
-
-const testRelation3 = defineRelation({
-  name: "department_company",
-  from: "department",
-  to: "company",
-  cardinality: "many_to_one",
-  fromName: "company",
-  toName: "departments",
-});
-
 const testView = defineView({
   name: "order_list",
   entity: "order",
@@ -1136,198 +1118,5 @@ describe("formatFieldDesc", () => {
     });
     expect(result).toContain("minLength: 3");
     expect(result).toContain("maxLength: 10");
-  });
-});
-
-// ── Replicate relations helpers from describe-formatters.ts ──
-
-interface RelationSummary {
-  name: string;
-  from: string;
-  to: string;
-  cardinality: string;
-  fromName: string;
-  toName: string;
-}
-
-interface RelationsOverview {
-  total: number;
-  relations: RelationSummary[];
-  bySourceEntity: Record<string, RelationSummary[]>;
-}
-
-function buildRelationsOverview(relations: RelationDefinition[]): RelationsOverview {
-  const summaries: RelationSummary[] = relations.map((r) => ({
-    name: r.name,
-    from: r.from,
-    to: r.to,
-    cardinality: r.cardinality,
-    fromName: r.fromName,
-    toName: r.toName,
-  }));
-
-  const bySourceEntity: Record<string, RelationSummary[]> = {};
-  for (const s of summaries) {
-    if (!bySourceEntity[s.from]) {
-      bySourceEntity[s.from] = [];
-    }
-    bySourceEntity[s.from].push(s);
-  }
-
-  return { total: relations.length, relations: summaries, bySourceEntity };
-}
-
-function printRelationsOverview(overview: RelationsOverview): void {
-  console.log("");
-  console.log("  Relation Graph Overview");
-  console.log("  =======================");
-  console.log("");
-  console.log(`  Total relations: ${overview.total}`);
-  console.log("");
-
-  if (overview.total === 0) {
-    console.log("  (no relations defined)");
-    console.log("");
-    return;
-  }
-
-  const sourceEntities = Object.keys(overview.bySourceEntity).sort();
-  for (const entity of sourceEntities) {
-    const relations = overview.bySourceEntity[entity];
-    console.log(`  ${entity}:`);
-    for (const r of relations) {
-      console.log(
-        `    - ${r.name}: ${r.from} -> ${r.to} (${r.cardinality}) [${r.fromName} / ${r.toName}]`,
-      );
-    }
-    console.log("");
-  }
-}
-
-// ── buildRelationsOverview ──────────────────────────────
-
-describe("buildRelationsOverview", () => {
-  test("returns correct total count", () => {
-    const overview = buildRelationsOverview([testRelation, testRelation2, testRelation3]);
-    expect(overview.total).toBe(3);
-  });
-
-  test("returns empty overview for no relations", () => {
-    const overview = buildRelationsOverview([]);
-    expect(overview.total).toBe(0);
-    expect(overview.relations).toHaveLength(0);
-    expect(Object.keys(overview.bySourceEntity)).toHaveLength(0);
-  });
-
-  test("maps relation fields correctly", () => {
-    const overview = buildRelationsOverview([testRelation]);
-    expect(overview.relations).toHaveLength(1);
-    const r = overview.relations[0];
-    expect(r.name).toBe("order_department");
-    expect(r.from).toBe("order");
-    expect(r.to).toBe("department");
-    expect(r.cardinality).toBe("many_to_one");
-    expect(r.fromName).toBe("department");
-    expect(r.toName).toBe("orders");
-  });
-
-  test("groups relations by source entity", () => {
-    const overview = buildRelationsOverview([testRelation, testRelation2, testRelation3]);
-    expect(Object.keys(overview.bySourceEntity)).toHaveLength(2);
-    // order has 2 relations, department has 1
-    expect(overview.bySourceEntity.order).toHaveLength(2);
-    expect(overview.bySourceEntity.department).toHaveLength(1);
-  });
-
-  test("single relation grouped under its source", () => {
-    const overview = buildRelationsOverview([testRelation3]);
-    expect(overview.bySourceEntity.department).toHaveLength(1);
-    expect(overview.bySourceEntity.department[0].name).toBe("department_company");
-  });
-});
-
-// ── printRelationsOverview ──────────────────────────────
-
-describe("printRelationsOverview", () => {
-  test("prints header and total count", () => {
-    const overview = buildRelationsOverview([testRelation]);
-    printRelationsOverview(overview);
-    const out = getOutput();
-    expect(out).toContain("Relation Graph Overview");
-    expect(out).toContain("=======================");
-    expect(out).toContain("Total relations: 1");
-  });
-
-  test("prints empty message when no relations", () => {
-    const overview = buildRelationsOverview([]);
-    printRelationsOverview(overview);
-    const out = getOutput();
-    expect(out).toContain("Total relations: 0");
-    expect(out).toContain("(no relations defined)");
-  });
-
-  test("prints relations grouped by source entity", () => {
-    const overview = buildRelationsOverview([testRelation, testRelation2, testRelation3]);
-    printRelationsOverview(overview);
-    const out = getOutput();
-    // Source entities appear as section headers
-    expect(out).toContain("  order:");
-    expect(out).toContain("  department:");
-  });
-
-  test("prints relation details with cardinality and semantic names", () => {
-    const overview = buildRelationsOverview([testRelation]);
-    printRelationsOverview(overview);
-    const out = getOutput();
-    expect(out).toContain(
-      "order_department: order -> department (many_to_one) [department / orders]",
-    );
-  });
-
-  test("sorts source entities alphabetically", () => {
-    const overview = buildRelationsOverview([testRelation, testRelation3]);
-    printRelationsOverview(overview);
-    const out = getOutput();
-    const departmentIdx = out.indexOf("  department:");
-    const orderIdx = out.indexOf("  order:");
-    expect(departmentIdx).toBeLessThan(orderIdx);
-  });
-
-  test("shows multiple relations under same source entity", () => {
-    const overview = buildRelationsOverview([testRelation, testRelation2]);
-    printRelationsOverview(overview);
-    const out = getOutput();
-    expect(out).toContain("order_department: order -> department");
-    expect(out).toContain("order_customer: order -> customer");
-  });
-});
-
-// ── Relations JSON output ───────────────────────────────
-
-describe("relations JSON output", () => {
-  test("buildRelationsOverview returns serializable JSON structure", () => {
-    const overview = buildRelationsOverview([testRelation, testRelation2]);
-    const json = JSON.stringify(overview, null, 2);
-    const parsed = JSON.parse(json);
-    expect(parsed.total).toBe(2);
-    expect(parsed.relations).toHaveLength(2);
-    expect(parsed.bySourceEntity.order).toHaveLength(2);
-  });
-
-  test("JSON includes fromName and toName for each relation", () => {
-    const overview = buildRelationsOverview([testRelation]);
-    const json = JSON.stringify(overview, null, 2);
-    const parsed = JSON.parse(json);
-    expect(parsed.relations[0].fromName).toBe("department");
-    expect(parsed.relations[0].toName).toBe("orders");
-  });
-
-  test("JSON handles empty relations", () => {
-    const overview = buildRelationsOverview([]);
-    const json = JSON.stringify(overview, null, 2);
-    const parsed = JSON.parse(json);
-    expect(parsed.total).toBe(0);
-    expect(parsed.relations).toEqual([]);
-    expect(parsed.bySourceEntity).toEqual({});
   });
 });
